requirements.txt
# Only required non-stdlib dependency for core functionality
Pillow>=9.0.0
